requirements.txt
# Core AI Framework
agno>=0.1.0

# Web Framework
streamlit>=1.52.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=12.0.0

# Database and Storage
lancedb>=0.3.0

# AI and ML
openai>=1.0.0

# Image Processing
Pillow>=10.0.0
pydicom>=2.4.0

# Document Processing
python-docx>=0.8.11
pypdf>=3.15.0
beautifulsoup4>=4.12.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
rich>=13.0.0
